+++ Tests/Token/FunctionTest.php
@@ -2,7 +2,7 @@
 /**
  * php-token-stream
  *
- * Copyright (c) 2009-2011, Sebastian Bergmann <sb@sebastian-bergmann.de>.
+ * Copyright (c) 2009-2012, Sebastian Bergmann <sb@sebastian-bergmann.de>.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -37,7 +37,7 @@
  * @package    PHP_TokenStream
  * @subpackage Tests
  * @author     Sebastian Bergmann <sb@sebastian-bergmann.de>
- * @copyright  2009-2011 Sebastian Bergmann <sb@sebastian-bergmann.de>
+ * @copyright  2009-2012 Sebastian Bergmann <sb@sebastian-bergmann.de>
  * @license    http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @since      File available since Release 1.0.0
  */
@@ -58,7 +58,7 @@ require_once 'PHP/Token/Stream.php';
  * @package    PHP_TokenStream
  * @subpackage Tests
  * @author     Sebastian Bergmann <sb@sebastian-bergmann.de>
- * @copyright  2009-2011 Sebastian Bergmann <sb@sebastian-bergmann.de>
+ * @copyright  2009-2012 Sebastian Bergmann <sb@sebastian-bergmann.de>
  * @license    http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @version    Release: @package_version@
  * @link       http://github.com/sebastianbergmann/php-token-stream/
@@ -85,9 +85,19 @@ class PHP_Token_FunctionTest extends PHPUnit_Framework_TestCase
     public function testGetArguments()
     {
         $this->assertEquals(array(), $this->functions[0]->getArguments());
-        $this->assertEquals(array('$baz' => 'Baz'), $this->functions[1]->getArguments());
-        $this->assertEquals(array('$foobar' => 'Foobar'), $this->functions[2]->getArguments());
-        $this->assertEquals(array('$barfoo' => 'Barfoo'), $this->functions[3]->getArguments());
+
+        $this->assertEquals(
+          array('$baz' => 'Baz'), $this->functions[1]->getArguments()
+        );
+
+        $this->assertEquals(
+          array('$foobar' => 'Foobar'), $this->functions[2]->getArguments()
+        );
+
+        $this->assertEquals(
+          array('$barfoo' => 'Barfoo'), $this->functions[3]->getArguments()
+        );
+
         $this->assertEquals(array(), $this->functions[4]->getArguments());
     }
 
@@ -133,8 +143,17 @@ class PHP_Token_FunctionTest extends PHPUnit_Framework_TestCase
     public function testGetDocblock()
     {
         $this->assertNull($this->functions[0]->getDocblock());
-        $this->assertEquals("/**\n     * @param Baz \$baz\n     */", $this->functions[1]->getDocblock());
-        $this->assertEquals("/**\n     * @param Foobar \$foobar\n     */", $this->functions[2]->getDocblock());
+
+        $this->assertEquals(
+          "/**\n     * @param Baz \$baz\n     */",
+          $this->functions[1]->getDocblock()
+        );
+
+        $this->assertEquals(
+          "/**\n     * @param Foobar \$foobar\n     */",
+          $this->functions[2]->getDocblock()
+        );
+
         $this->assertNull($this->functions[3]->getDocblock());
         $this->assertNull($this->functions[4]->getDocblock());
     }