+++ Tests/_files/classInNamespace.php
@@ -1,8 +1,6 @@
 <?php
-
 namespace Foo\Bar;
 
-class TestClass {
-
+class TestClass
+{
 }
-