+++ Tests/_files/multipleNamespacesWithOneClassUsingBraces.php
@@ -1,14 +1,12 @@
 <?php
-
 namespace Foo\Bar;
 
-class TestClassInBar {
-
+class TestClassInBar
+{
 }
 
 namespace Foo\Baz;
 
-class TestClassInBaz {
-
+class TestClassInBaz
+{
 }
-