+++ ChangeLog.markdown
@@ -3,6 +3,17 @@ PHP_TokenStream 1.1
 
 This is the list of changes for the PHP_TokenStream 1.1 release series.
 
+PHP_TokenStream 1.1.2
+---------------------
+
+* Added classes for the `T_CALLABLE` and `T_INSTEADOF` tokens that were added in PHP 5.4.
+* Added support for namespaced functions.
+
+PHP_TokenStream 1.1.1
+---------------------
+
+* Fixed issue #19: Notice in `PHP_Token_INTERFACE::hasInterfaces()`.
+
 PHP_TokenStream 1.1.0
 ---------------------